\name{nlsModel}
\alias{nlsModel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 calculate a model as defined in nls for the set of parameters \code{start} on the values \code{data}
}
\description{
Recovered from older versions of R (<= 1.8), obsolete in 2.10
}
\usage{
nlsModel(form, data, start)
}
\arguments{
  \item{form}{
A nls model object
}
  \item{data}{
 a data frame of independent variables
}
  \item{start}{
 a list of parameter values
}
}
\value{
a vector of estimates
}

\author{
R development core team
}
\keyword{ nonlinear }
